Place all the fixture related test support code in its own file.
