+++ tests/test_mig_shared_configuration.py
@@ -31,8 +31,9 @@ import inspect
 import os
 import unittest
 
-from tests.support import MigTestCase, TEST_DATA_DIR, PY2, testmain, \
-    fixturefile
+from tests.support import MigTestCase, TEST_DATA_DIR, PY2, testmain
+from tests.support.fixturesupp import FixtureAssertMixin
+
 from mig.shared.configuration import Configuration
 
 
@@ -45,7 +46,7 @@ def _to_dict(obj):
             if not (k.startswith('__') or _is_method(v))}
 
 
-class MigSharedConfiguration(MigTestCase):
+class MigSharedConfiguration(MigTestCase, FixtureAssertMixin):
     """Wrap unit tests for the corresponding module"""
 
     def test_argument_storage_protocols(self):
